pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ha-controld-manager"
version = "0.1.0"
description = "Home Assistant custom integration scaffold for Control D Manager"
readme = "README.md"
requires-python = ">=3.14"
dependencies = []

[project.optional-dependencies]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-homeassistant-custom-component>=0.13.218",
    "ruff>=0.11.0",
]

[tool.setuptools.packages.find]
include = ["custom_components*"]

[tool.pytest.ini_options]
addopts = "-ra"
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]
filterwarnings = [
    "ignore:Setting custom ClientSession.close attribute is discouraged:DeprecationWarning:homeassistant.helpers.aiohttp_client",
]

[tool.ruff]
line-length = 88
target-version = "py314"

[tool.ruff.lint]
select = ["ASYNC", "B", "C4", "E", "F", "I", "RUF", "SIM", "UP"]

[tool.mypy]
python_version = "3.14"
strict = true
warn_unused_ignores = true
show_error_codes = true
files = ["custom_components/controld_manager"]

[tool.pylint.main]
ignore-paths = [
    "^\\.mypy_cache/.*$",
    "^\\.pytest_cache/.*$",
    "^\\.ruff_cache/.*$",
    "^\\.tmp/.*$",
]

[tool.pylint."messages control"]
disable = ["line-too-long"]